+++ src/Model/DiscountList.php
@@ -7,17 +7,21 @@
 
 namespace OxidSolutionCatalysts\Unzer\Model;
 
+use OxidEsales\Eshop\Application\Model\Article;
+use OxidEsales\Eshop\Application\Model\Groups;
+use OxidEsales\Eshop\Application\Model\User;
+use OxidEsales\Eshop\Core\DatabaseProvider;
+use OxidEsales\Eshop\Core\Model\ListModel;
 use OxidEsales\Eshop\Core\Registry;
+use OxidEsales\Eshop\Application\Model\DiscountList as CoreDisCountList;
 
 class DiscountList extends DiscountList_parent
 {
-
-
     /**
      * Returns array of discounts that can be globally (transparently) applied
      *
-     * @param \OxidEsales\Eshop\Application\Model\Article $oArticle article object
-     * @param \OxidEsales\Eshop\Application\Model\User    $oUser    oxuser object (optional)
+     * @param Article $oArticle article object
+     * @param User    $oUser    oxuser object (optional)
      *
      * @return array
      */
@@ -25,7 +29,9 @@ class DiscountList extends DiscountList_parent
     {
         $aList = [];
         $this->forceReload();
-        $aDiscList = $this->_getList($oUser)->getArray();
+        /** @var CoreDisCountList $oDiscList */
+        $oDiscList = $this->_getList($oUser);
+        $aDiscList = $oDiscList->getArray();
         foreach ($aDiscList as $oDiscount) {
             if ($oDiscount->isForArticle($oArticle)) {
                 $aList[$oDiscount->getId()] = $oDiscount;
@@ -39,33 +45,35 @@ class DiscountList extends DiscountList_parent
     /**
      * Creates discount list filter SQL to load current state discount list
      *
-     * @param \OxidEsales\Eshop\Application\Model\User $oUser user object
+     * @SuppressWarnings(PHPMD.StaticAccess)
      *
-     * @return string
-     * @deprecated underscore prefix violates PSR12, will be renamed to "getFilterSelect" in next major
+     * @inheritdoc
      */
-    protected function _getFilterSelect($oUser) // phpcs:ignore PSR2.Methods.MethodDeclaration.Underscore
+    protected function _getFilterSelect($oUser)
     {
         $oBaseObject = $this->getBaseObject();
-        $bDisableSqlActiveSnippet = Registry::getSession()->getVariable('disableSqlActiveSnippet');
+        $bDisableSqlActive = Registry::getSession()->getVariable('disableSqlActiveSnippet');
         $sTable = $oBaseObject->getViewName();
-        $sQ = "select " . $oBaseObject->getSelectFields() . " from $sTable ";
-        $sQ .= true === $bDisableSqlActiveSnippet ? "where 1=1 " : "where " . $oBaseObject->getSqlActiveSnippet() . ' ';
-
+        $sSql = "select " . $oBaseObject->getSelectFields() . " from $sTable  where 1 ";
+        $sSql .= false === $bDisableSqlActive ? $oBaseObject->getSqlActiveSnippet() . ' ' : '';
 
         // defining initial filter parameters
         $sUserId = null;
         $sGroupIds = null;
         $sCountryId = $this->getCountryId($oUser);
-        $oDb = \OxidEsales\Eshop\Core\DatabaseProvider::getDb();
+        $oDb = DatabaseProvider::getDb();
 
-        // checking for current session user which gives additional restrictions for user itself, users group and country
+        // checking for current session user which gives additional restrictions for user itself,
+        // users group and country
         if ($oUser) {
             // user ID
             $sUserId = $oUser->getId();
 
             // user group ids
-            foreach ($oUser->getUserGroups() as $oGroup) {
+            /** @var ListModel $userGroups */
+            $userGroups = $oUser->getUserGroups();
+            /** @var Groups $oGroup */
+            foreach ($userGroups as $oGroup) {
                 if ($sGroupIds) {
                     $sGroupIds .= ', ';
                 }
@@ -77,24 +85,45 @@ class DiscountList extends DiscountList_parent
         $sGroupTable = getViewName('oxgroups');
         $sCountryTable = getViewName('oxcountry');
 
-        $sCountrySql = $sCountryId ? "EXISTS(select oxobject2discount.oxid from oxobject2discount where oxobject2discount.OXDISCOUNTID=$sTable.OXID and oxobject2discount.oxtype='oxcountry' and oxobject2discount.OXOBJECTID=" . $oDb->quote($sCountryId) . ")" : '0';
-        $sUserSql = $sUserId ? "EXISTS(select oxobject2discount.oxid from oxobject2discount where oxobject2discount.OXDISCOUNTID=$sTable.OXID and oxobject2discount.oxtype='oxuser' and oxobject2discount.OXOBJECTID=" . $oDb->quote($sUserId) . ")" : '0';
-        $sGroupSql = $sGroupIds ? "EXISTS(select oxobject2discount.oxid from oxobject2discount where oxobject2discount.OXDISCOUNTID=$sTable.OXID and oxobject2discount.oxtype='oxgroups' and oxobject2discount.OXOBJECTID in ($sGroupIds) )" : '0';
+        $sCountrySql = $sCountryId ?
+            "EXISTS(select oxobject2discount.oxid from oxobject2discount where
+            oxobject2discount.OXDISCOUNTID = $sTable.OXID and oxobject2discount.oxtype = 'oxcountry' and
+            oxobject2discount.OXOBJECTID = " . $oDb->quote($sCountryId) . ")" :
+            '0';
+        $sUserSql = $sUserId ?
+            "EXISTS(select oxo bject2discount.oxid from oxobject2discount where
+            oxobject2discount.OXDISCOUNTID = $sTable.OXID and oxobject2discount.oxtype = 'oxuser' and
+            oxobject2discount.OXOBJECTID = " . $oDb->quote($sUserId) . ")" :
+            '0';
+        $sGroupSql = $sGroupIds ?
+            "EXISTS(select oxobject2discount.oxid from oxobject2discount where
+            oxobject2discount.OXDISCOUNTID = $sTable.OXID and oxobject2discount.oxtype = 'oxgroups' and
+            oxobject2discount.OXOBJECTID in ($sGroupIds) )" :
+            '0';
 
-        $sQ .= "and (
-                if(EXISTS(select 1 from oxobject2discount, $sCountryTable where $sCountryTable.oxid=oxobject2discount.oxobjectid and oxobject2discount.OXDISCOUNTID=$sTable.OXID and oxobject2discount.oxtype='oxcountry' LIMIT 1),
+        $sSql .= "and (
+                if(EXISTS(select 1 from oxobject2discount, $sCountryTable where
+                $sCountryTable.oxid = oxobject2discount.oxobjectid and
+                oxobject2discount.OXDISCOUNTID = $sTable.OXID and
+                oxobject2discount.oxtype = 'oxcountry' LIMIT 1),
                         $sCountrySql,
                         1) &&
-                if(EXISTS(select 1 from oxobject2discount, $sUserTable where $sUserTable.oxid=oxobject2discount.oxobjectid and oxobject2discount.OXDISCOUNTID=$sTable.OXID and oxobject2discount.oxtype='oxuser' LIMIT 1),
+                if(EXISTS(select 1 from oxobject2discount, $sUserTable where
+                $sUserTable.oxid = oxobject2discount.oxobjectid and
+                oxobject2discount.OXDISCOUNTID = $sTable.OXID and
+                oxobject2discount.oxtype = 'oxuser' LIMIT 1),
                         $sUserSql,
                         1) &&
-                if(EXISTS(select 1 from oxobject2discount, $sGroupTable where $sGroupTable.oxid=oxobject2discount.oxobjectid and oxobject2discount.OXDISCOUNTID=$sTable.OXID and oxobject2discount.oxtype='oxgroups' LIMIT 1),
+                if(EXISTS(select 1 from oxobject2discount, $sGroupTable where 
+                $sGroupTable.oxid = oxobject2discount.oxobjectid and 
+                oxobject2discount.OXDISCOUNTID = $sTable.OXID and
+                oxobject2discount.oxtype = 'oxgroups' LIMIT 1),
                         $sGroupSql,
                         1)
             )";
 
-        $sQ .= " order by $sTable.oxsort ";
+        $sSql .= " order by $sTable.oxsort ";
 
-        return $sQ;
+        return $sSql;
     }
 }